+++ calculator/calculator_test.go
@@ -1,4 +1,4 @@
-package main
+package calculator
 
 import (
 	"testing"
