+++ model/calculation.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 type Calculation struct {
 	typeOfCalculation string
